refactor: screen 파일 이동
