+++ app/src/main/java/com/fintern/ourbudgeting/ui/assetmanagement/assetdisplay/AssetDisplayScreen.kt
@@ -1,9 +1,10 @@
-package com.fintern.ourbudgeting.ui.assetmanagement.assetedition.component
+package com.fintern.ourbudgeting.ui.assetmanagement.assetdisplay
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -11,19 +12,22 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.fintern.ourbudgeting.R
+import com.fintern.ourbudgeting.ui.assetmanagement.assetdisplay.component.AssetDisplayTopAppBar
+import com.fintern.ourbudgeting.ui.assetmanagement.common.component.AssetBody
+import com.fintern.ourbudgeting.ui.assetmanagement.common.component.AssetDivider
 import com.fintern.ourbudgeting.ui.assetmanagement.common.component.AssetTitle
 import com.fintern.ourbudgeting.ui.assetmanagement.common.component.AssetTopSection
-import com.fintern.ourbudgeting.ui.theme.OurBudgetingTheme
 
 @Composable
-fun AssetEditScreen(
+fun AssetDisplayScreen(
     modifier: Modifier = Modifier,
     asset: Long,
-    debt: Long
+    dept: Long,
 ) {
     Scaffold(
-        modifier = modifier.background(Color.White),
-        topBar = { AssetEditTopAppBar() }
+        modifier = modifier,
+        topBar = { AssetDisplayTopAppBar() },
+        floatingActionButton = { }
     ) { paddingValue ->
         Column(
             modifier = Modifier
@@ -31,19 +35,20 @@ fun AssetEditScreen(
                 .background(Color.White)
                 .padding(paddingValue)
         ) {
-            AssetTopSection(asset = asset, debt = debt)
+            AssetTopSection(asset = asset, debt = dept)
+            AssetDivider()
             AssetTitle(name = stringResource(R.string.cash), amount = 1000)
-            AssetEditBody(name = stringResource(R.string.cash), amount = 1000)
+            AssetBody(name = stringResource(R.string.cash), amount = 1000)
             AssetTitle(name = stringResource(R.string.card), amount = 2000)
-            AssetEditBody(name = stringResource(R.string.card), amount = 1000)
+            AssetBody(name = stringResource(R.string.card), amount = 1000)
         }
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun ModifyAssetManagementScreenPreview() {
-    OurBudgetingTheme {
-        AssetEditScreen(asset = 300, debt = 2000)
+fun AssetManagementScreenPreview() {
+    MaterialTheme {
+        AssetDisplayScreen(asset = 1000, dept = 2000)
     }
 }
