+++ app/src/main/java/com/fintern/ourbudgeting/ui/assetmanagement/assetedition/AssetEditScreen.kt
@@ -1,10 +1,9 @@
-package com.fintern.ourbudgeting.ui.assetmanagement.assetdisplay.component
+package com.fintern.ourbudgeting.ui.assetmanagement.assetedition
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -12,21 +11,21 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.fintern.ourbudgeting.R
-import com.fintern.ourbudgeting.ui.assetmanagement.common.component.AssetBody
-import com.fintern.ourbudgeting.ui.assetmanagement.common.component.AssetDivider
+import com.fintern.ourbudgeting.ui.assetmanagement.assetedition.component.AssetEditBody
+import com.fintern.ourbudgeting.ui.assetmanagement.assetedition.component.AssetEditTopAppBar
 import com.fintern.ourbudgeting.ui.assetmanagement.common.component.AssetTitle
 import com.fintern.ourbudgeting.ui.assetmanagement.common.component.AssetTopSection
+import com.fintern.ourbudgeting.ui.theme.OurBudgetingTheme
 
 @Composable
-fun AssetDisplayScreen(
+fun AssetEditScreen(
     modifier: Modifier = Modifier,
     asset: Long,
-    dept: Long,
+    debt: Long
 ) {
     Scaffold(
-        modifier = modifier,
-        topBar = { AssetDisplayTopAppBar() },
-        floatingActionButton = { }
+        modifier = modifier.background(Color.White),
+        topBar = { AssetEditTopAppBar() }
     ) { paddingValue ->
         Column(
             modifier = Modifier
@@ -34,20 +33,19 @@ fun AssetDisplayScreen(
                 .background(Color.White)
                 .padding(paddingValue)
         ) {
-            AssetTopSection(asset = asset, debt = dept)
-            AssetDivider()
+            AssetTopSection(asset = asset, debt = debt)
             AssetTitle(name = stringResource(R.string.cash), amount = 1000)
-            AssetBody(name = stringResource(R.string.cash), amount = 1000)
+            AssetEditBody(name = stringResource(R.string.cash), amount = 1000)
             AssetTitle(name = stringResource(R.string.card), amount = 2000)
-            AssetBody(name = stringResource(R.string.card), amount = 1000)
+            AssetEditBody(name = stringResource(R.string.card), amount = 1000)
         }
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun AssetManagementScreenPreview() {
-    MaterialTheme {
-        AssetDisplayScreen(asset = 1000, dept = 2000)
+fun ModifyAssetManagementScreenPreview() {
+    OurBudgetingTheme {
+        AssetEditScreen(asset = 300, debt = 2000)
     }
 }
