+++ app/src/main/java/com/fintern/ourbudgeting/ui/assetmanagement/assettypeaddition/AssetAdditionScreen.kt
@@ -1,4 +1,4 @@
-package com.fintern.ourbudgeting.ui.assetmanagement.assettypeaddition.component
+package com.fintern.ourbudgeting.ui.assetmanagement.assettypeaddition
 
 import android.widget.Toast
 import androidx.compose.foundation.background
@@ -16,7 +16,9 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.fintern.ourbudgeting.R
-import com.fintern.ourbudgeting.ui.assetmanagement.assettypeaddition.AssetAdditionViewModel
+import com.fintern.ourbudgeting.ui.assetmanagement.assettypeaddition.component.AssetAdditionBody
+import com.fintern.ourbudgeting.ui.assetmanagement.assettypeaddition.component.AssetAdditionButton
+import com.fintern.ourbudgeting.ui.assetmanagement.assettypeaddition.component.AssetAdditionTopAppBar
 
 @Composable
 fun AssetAdditionScreen(
